Move shaders to own files to reuse
