+++ tests/visual/scenes/filters/clip-top-viewport/circle.wgsl
@@ -1,0 +1,89 @@
+
+struct GlobalFilterUniforms {
+  uInputSize:vec4<f32>,
+  uInputPixel:vec4<f32>,
+  uInputClamp:vec4<f32>,
+  uOutputFrame:vec4<f32>,
+  uGlobalFrame:vec4<f32>,
+  uOutputTexture:vec4<f32>,
+};
+
+@group(0) @binding(0) var<uniform> gfu: GlobalFilterUniforms;
+@group(0) @binding(1) var uTexture: texture_2d<f32>;
+@group(0) @binding(2) var uSampler: sampler;
+
+const center: vec2<f32> = vec2(0.5);
+const angleToRotate: f32 = 180.0 * (3.14159265 / 180.0);
+
+fn mapCoord(coord: vec2<f32> ) -> vec2<f32>
+{
+  var mappedCoord: vec2<f32> = coord;
+  mappedCoord *= gfu.uInputSize.xy;
+  mappedCoord += gfu.uOutputFrame.xy;
+  return mappedCoord;
+}
+
+fn unmapCoord(coord: vec2<f32> ) -> vec2<f32>
+{
+  var mappedCoord: vec2<f32> = coord;
+  mappedCoord -= gfu.uOutputFrame.xy;
+  mappedCoord /= gfu.uInputSize.xy;
+  return mappedCoord;
+}
+
+@fragment
+fn mainFragment(
+    @builtin(position) position: vec4<f32>,
+    @location(0) uv : vec2<f32>
+) -> @location(0) vec4<f32>
+{
+    var centerInPx: vec2<f32> = mapCoord(center);
+    var coord: vec2<f32> = mapCoord(uv);
+    coord -= centerInPx;
+
+    var radius: f32 = length(coord);
+    var angle: f32 = atan2(coord.y, coord.x) + angleToRotate;
+
+    var rotatedPoint: vec2<f32> = coord;
+
+    if (radius < 32.0) {
+        rotatedPoint = vec2(cos(angle), sin(angle)) * radius;
+    }
+
+    rotatedPoint += centerInPx;
+    rotatedPoint = unmapCoord(rotatedPoint);
+
+    var resultColor: vec4<f32> = textureSample(uTexture, uSampler, rotatedPoint);
+
+    return resultColor;
+}
+
+struct VSOutput {
+    @builtin(position) position: vec4<f32>,
+    @location(0) uv : vec2<f32>
+  };
+
+fn filterVertexPosition(aPosition:vec2<f32>) -> vec4<f32>
+{
+    var position = aPosition * gfu.uOutputFrame.zw + gfu.uOutputFrame.xy;
+
+    position.x = position.x * (2.0 / gfu.uOutputTexture.x) - 1.0;
+    position.y = position.y * (2.0*gfu.uOutputTexture.z / gfu.uOutputTexture.y) - gfu.uOutputTexture.z;
+
+    return vec4(position, 0.0, 1.0);
+}
+
+fn filterTextureCoord( aPosition:vec2<f32> ) -> vec2<f32>
+{
+    return aPosition * (gfu.uOutputFrame.zw * gfu.uInputSize.zw);
+}
+
+@vertex
+fn mainVertex(
+  @location(0) aPosition : vec2<f32>,
+) -> VSOutput {
+  return VSOutput(
+   filterVertexPosition(aPosition),
+   filterTextureCoord(aPosition)
+  );
+}
